user-betting service with tests

+++ common/src/main/java/f1/betting/poc/domain/Driver.java
@@ -7,7 +7,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Driver {
-	private int driverNumber;
+	private Long driverNumber;
 	private String fullName;
 	private String teamName;
 	/**
